[no ci] some modifications pyddff
