+++ ddff/pyddff/pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "pyddff"
-version = "1.0"
+version = "1.0.2"
 description = "Pure python interface to reading ddff files"
 readme = "README.md"
 requires-python = ">=3.6"
@@ -13,6 +13,9 @@ classifiers = [
 
 license = "EUPL-1.2"
 
+[project.scripts]
+ddff-convert = "pyddff.ddff_convert:main"
+
 [dependencies]
 msgpack-python = "^1.0.3"
 crcmod = "1.7"
@@ -20,6 +23,7 @@ crcmod = "1.7"
 [build-system]
 requires = [
     "setuptools >= 44",
-    "wheel"
+    "wheel",
+    "h5py"
 ]
 build-backend = "setuptools.build_meta"
