+++ ddff/pyddff/src/__init__.py
@@ -1,3 +1,4 @@
 from .ddffbase import DDFF, DDFFStream, vprint
 from .ddffinventoried import DDFFInventoriedStream, DDFFInventoried
 from .ddfftagged import DDFFTagStream, DDFFTagged
+from .ddff_convert import main
